+++ windows/src/universal_ble_plugin.h
@@ -125,8 +125,8 @@ namespace universal_ble
                                           const std::vector<uint8_t> &value,
                                           std::function<void(std::optional<FlutterError> reply)> result);
         bool filterByManufacturerData(IVector<BluetoothLEManufacturerData> deviceManufactureData);
-        winrt::fire_and_forget PairAsync(std::string device_id, const DeviceInformation deviceInformation);
-        winrt::fire_and_forget CustomPairAsync(std::string device_id, const DeviceInformation deviceInformation);
+        winrt::fire_and_forget PairAsync(std::string device_id, std::function<void(ErrorOr<bool> reply)> result);
+        winrt::fire_and_forget CustomPairAsync(std::string device_id, std::function<void(ErrorOr<bool> reply)> result);
         void PairingRequestedHandler(DeviceInformationCustomPairing sender, DevicePairingRequestedEventArgs eventArgs);
 
         // UniversalBlePlatformChannel implementation.
@@ -165,7 +165,9 @@ namespace universal_ble
         void IsPaired(
             const std::string &device_id,
             std::function<void(ErrorOr<bool> reply)> result) override;
-        std::optional<FlutterError> Pair(const std::string &device_id) override;
+        void Pair(
+            const std::string &device_id,
+            std::function<void(ErrorOr<bool> reply)> result) override;
         std::optional<FlutterError> UnPair(const std::string &device_id) override;
         void GetSystemDevices(
             const flutter::EncodableList &with_services,